refactor: Update TechStack component with Tailwind CSS logo
This commit updates the TechStack component by replacing the previous Tailwind CSS logo with a new SVG logo. The new logo is more visually appealing and aligns better with the overall design of the component.

The previous logo was a JPEG image, but the new logo is an SVG file, which allows for better scalability and flexibility. This change enhances the visual presentation of the TechStack component and improves the user experience.

The updated logo file can be found at src/assets/Images/tailwind-css-logo.svg.

Refactor the TechStack component to use the new logo file and remove the old logo file.